Use parcel as a bundler
