--- .gitattributes
@@ -1,2 +1,0 @@
-* linguist-vendored
-*.js linguist-vendored=false
