+++ .gitignore
@@ -1,0 +1,4 @@
+.cache
+.idea
+dist
+node_modules
