+++ CONTRIBUTING.md
@@ -1,6 +1,9 @@
 # Contributing to Death to IE11
 
 ## Adding a New Reason or website
-1. Open [reasons.json](reasons.json) or [websites.json](websites.json)
+1. Open [src/data/reasons.js](/src/data/reasons.js) or [src/data/websites.js](/src/data/websites.js)
 2. inside "list", add an entry with the following : "url", "label", "description"(optional)
 3. Commit your changes and submit a pull request
+
+## Other types of contributions
+If you find a way to improve the codebase or find bugs feel free to open an issue or to contribute, all the source code is in the `src` directory.
