+++ README.md
@@ -1,3 +1,4 @@
+[![Netlify Status](https://api.netlify.com/api/v1/badges/34c79567-f40e-4b57-ab19-8d1765cae4eb/deploy-status)](https://app.netlify.com/sites/death-to-ie11/deploys)
 # Death to Internet Explorer 11
 
 this is a fun little app that counts down to the end of support of Internet Explorer 11 + a few more features like a list of articles/websites that tell you why you should stop supporting IE11 and a list of websites that already dropped support.
@@ -6,3 +7,10 @@ this is a fun little app that counts down to the end of support of Internet Expl
 
 Want to contribute by adding a few reasons why people should stop supporting IE11 or by adding websites that stopped supporting IE11?
 [How to contribute](CONTRIBUTING.md)
+
+## Running Locally
+Install dependencies: `yarn install`
+
+Run the project: `yarn dev`
+
+navigate to the `src` directory and have fun! :)
