+++ src/index.js
@@ -1,0 +1,5 @@
+import countdown from "./js/countdown";
+import populateReasonsAndWebsites from './js/reasonsAndWebsites';
+
+countdown('10/14/2025 05:00:00 PM');
+populateReasonsAndWebsites();
